Do our own parsing rather than evalling.
This make things much more robust, but makes things way more complex.
Ths only thing we're presently lacking is the ability to expand escape
sequences. For example, we can't turn foo="\"" into 'foo="', so we
return the escapes.

Otherwise, this makes it safe(r?) to use on untrusted inputs, as we no
longer expand parameters (including command expansions) or globs or
anything.