Only calculate the "next" char if we're on a \.
This ends up having a significant performance impact. See, by supporting
older shells, we're kinda stuck in this position where we're calling
`expr` at least once for every character in our cmdline. This involves
a subshell (a fork/clone) which then calls `expr` (a fork/clone).
Currently, we use `expr` to grab the current character AND the next
character (ya know, just in case we're sitting on a \). For a 300-char
cmdline, this ends up taking around 1.3 seconds to find something at the
end.

We can cut the number of `expr` calls roughly in half if we only look at
the next character when we know we need it. This gives us a 40-50%
speedup.